refactor: Rename Website Builder to Builder
