+++ .github/workflows/ci.yml
@@ -8,7 +8,7 @@ on:
   pull_request:
 
 concurrency:
-  group: develop-website_builder-${{ github.event.number }}
+  group: develop-builder-${{ github.event.number }}
   cancel-in-progress: true
 
 jobs:
@@ -85,10 +85,10 @@ jobs:
       - name: Install
         working-directory: /home/runner/frappe-bench
         run: |
-          bench get-app website_builder $GITHUB_WORKSPACE
+          bench get-app builder $GITHUB_WORKSPACE
           bench setup requirements --dev
           bench new-site --db-root-password root --admin-password admin test_site
-          bench --site test_site install-app website_builder
+          bench --site test_site install-app builder
           bench build
         env:
           CI: 'Yes'
@@ -97,6 +97,6 @@ jobs:
         working-directory: /home/runner/frappe-bench
         run: |
           bench --site test_site set-config allow_tests true
-          bench --site test_site run-tests --app website_builder
+          bench --site test_site run-tests --app builder
         env:
           TYPE: server
