+++ builder/builder/doctype/builder_component/builder_component.json
@@ -44,7 +44,7 @@
  "links": [],
  "modified": "2023-08-27 11:56:45.322261",
  "modified_by": "Administrator",
- "module": "Website Builder",
+ "module": "Builder",
  "name": "Builder Component",
  "naming_rule": "By fieldname",
  "owner": "Administrator",
