+++ builder/builder/doctype/builder_page/builder_page.json
@@ -128,7 +128,7 @@
  "links": [],
  "modified": "2023-09-28 15:51:19.458047",
  "modified_by": "Administrator",
- "module": "Website Builder",
+ "module": "Builder",
  "name": "Builder Page",
  "naming_rule": "By fieldname",
  "owner": "Administrator",
