+++ builder/builder/doctype/builder_page/patches/properly_extend_blocks_from_component.py
