+++ builder/builder/patches/rename_web_page_beta_to_builder_page.py
