+++ builder/builder/patches/rename_web_page_component_to_builder_component.py
