+++ builder/patches.txt
@@ -1,0 +1,8 @@
+[pre_model_sync]
+builder.builder.doctype.builder_page.patches.create_upload_folder_for_builder #
+builder.builder.patches.rename_web_page_beta_to_builder_page
+builder.builder.patches.rename_web_page_component_to_builder_component
+
+[post_model_sync]
+builder.builder.doctype.builder_component.patches.set_component_id
+builder.builder.doctype.builder_page.patches.properly_extend_blocks_from_component
