+++ builder/templates/generators/webpage.html
@@ -5,8 +5,8 @@
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>{{title}}</title>
-	<link rel="icon" href="/assets/website_builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
-	<link rel="icon" href="/assets/website_builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
+	<link rel="icon" href="/assets/builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
+	<link rel="icon" href="/assets/builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	{% block meta_block %}
 		{% include "templates/includes/meta_block.html" %}
