+++ builder/www/builder.html
@@ -2,13 +2,13 @@
 <html lang="en">
 	<head>
 		<meta charset="UTF-8" />
-		<link rel="icon" href="/assets/website_builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
-		<link rel="icon" href="/assets/website_builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
+		<link rel="icon" href="/assets/builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
+		<link rel="icon" href="/assets/builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<!-- <script src="https://cdn.tailwindcss.com"></script> -->
 		<title>Frappe Builder</title>
-		<script type="module" crossorigin src="/assets/website_builder/frontend/assets/index.e448aea1.js"></script>
-		<link rel="stylesheet" href="/assets/website_builder/frontend/assets/index.e0d16a2e.css">
+		<script type="module" crossorigin src="/assets/builder/frontend/assets/index.e448aea1.js"></script>
+		<link rel="stylesheet" href="/assets/builder/frontend/assets/index.e0d16a2e.css">
 	</head>
 	<body class="overscroll-none dark:bg-zinc-800">
 		<div id="app"></div>
