+++ frontend/index.html
@@ -2,8 +2,8 @@
 <html lang="en">
 	<head>
 		<meta charset="UTF-8" />
-		<link rel="icon" href="/assets/website_builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
-		<link rel="icon" href="/assets/website_builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
+		<link rel="icon" href="/assets/builder/images/frappe_black.png"  media="(prefers-color-scheme: light)"/>
+		<link rel="icon" href="/assets/builder/images/frappe_white.png"  media="(prefers-color-scheme: dark)"/>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<!-- <script src="https://cdn.tailwindcss.com"></script> -->
 		<title>Frappe Builder</title>
