+++ frontend/package.json
@@ -5,8 +5,8 @@
   "name": "builder-ui",
   "scripts": {
     "dev": "vite",
-    "build": "vite build --base=/assets/website_builder/frontend/ && yarn copy-html-entry",
-    "copy-html-entry": "cp ../website_builder/public/frontend/index.html ../website_builder/www/builder.html",
+    "build": "vite build --base=/assets/builder/frontend/ && yarn copy-html-entry",
+    "copy-html-entry": "cp ../builder/public/frontend/index.html ../builder/www/builder.html",
     "preview": "vite preview"
   },
   "dependencies": {
