+++ frontend/src/pages/PagePreview.vue
@@ -147,7 +147,7 @@ const setPreviewURL = () => {
 		page: route.params.pageId,
 		...store.routeVariables,
 	};
-	previewRoute.value = `/api/method/website_builder.website_builder.doctype.builder_page.builder_page.get_page_preview_html?${Object.entries(
+	previewRoute.value = `/api/method/builder.builder.doctype.builder_page.builder_page.get_page_preview_html?${Object.entries(
 		queryParams
 	)
 		.map(([key, value]) => `${key}=${value}`)
