+++ setup.py
@@ -7,8 +7,8 @@ from setuptools.command.develop import develop
 with open("requirements.txt") as f:
 	install_requires = f.read().strip().split("\n")
 
-# get version from __version__ variable in website_builder/__init__.py
-from website_builder import __version__ as version
+# get version from __version__ variable in builder/__init__.py
+from builder import __version__ as version
 
 
 def install_playwright():
@@ -25,7 +25,7 @@ class RunDevelopCommand(develop):
 
 
 setup(
-	name="website_builder",
+	name="builder",
 	version=version,
 	author="Suraj Shetty",
 	author_email="surajshetty3416@gmail.com",
