--- website_builder/patches.txt
@@ -1,8 +1,0 @@
-[pre_model_sync]
-website_builder.website_builder.doctype.builder_page.patches.create_upload_folder_for_builder #
-website_builder.website_builder.patches.rename_web_page_beta_to_builder_page
-website_builder.website_builder.patches.rename_web_page_component_to_builder_component
-
-[post_model_sync]
-website_builder.website_builder.doctype.builder_component.patches.set_component_id
-website_builder.website_builder.doctype.builder_page.patches.properly_extend_blocks_from_component
